fix: 乱数生成アルゴリズムを本家に合わせるよう変更

+++ native/src/engine/match.cpp
@@ -278,7 +278,7 @@ int64_t PuyotanMatch::runBatch(int num_games, uint32_t seed) noexcept {
 int PuyotanMatch::nextInt(uint32_t& seed, int max) noexcept {
     assert(seed != 0u);
     seed ^= (seed << 13);
-    seed ^= (seed >> 17);
+    seed ^= static_cast<uint32_t>(static_cast<int32_t>(seed) >> 17);
     seed ^= (seed << 15);
     return static_cast<int>(seed % static_cast<uint32_t>(max));
 }
